Backend to Develop

See merge request s09-webmobile1-sub2/S09P12D207!60

+++ BE/src/main/java/com/danbi/domain/member/service/MemberService.java
@@ -43,8 +43,8 @@ public class MemberService {
                 .build();
 
         Point point = Point.builder()
-                .dewPoint((long)100)
-                .accumulateDewPoint((long)100)
+                .dewPoint(100L)
+                .accumulateDewPoint(100L)
                 .profile(profile).build();
         pointRepository.save(point);
